requirements.txt
Pillow~=9.1.0
numpy~=1.22.3
opencv-python~=4.5.5.64